fix(stability): round-3 bugs — syscalls, mutex, NARS, battery
Syscall layer (A1, A2/A4, A3, A5/D1, A7):
- A1: dedicated apollo-sysctl-worker thread + mpsc queue replaces per-call
  thread::spawn (eliminates thread-storm on sysctl fanout)
- A2/A4: zombie check via proc_taskinfo::is_zombie_pid before set_nice,
  FreezeProcess, UnfreezeProcess (PROC_PIDTBSDINFO, SZOMB=5)
- A3: FrozenEntry gains start_sec; unfreeze_pids_verified rejects PID
  recycled since freeze (A-B-A defense)
- A5/D1: capture jetsam priority before demote; restore original on unfreeze
  via memorystatus_control GET_PRIORITY_LIST (cmd 1)
- A7: mach_qos permanently_blocked now TTL-expires (60s) via
  permanently_blocked_since HashMap + mark_blocked helper

NARS & RL (B1, B3, B4, B5, B6):
- B1: NARS TruthValue confidence clamp 0.9999 -> 0.99 (remove dead zone)
- B3: overflow_guard dedup window 60s -> 8s for burst events
- B4: nars_decay_factor floor 0.80 -> 0.90 (no runaway forgetting)
- B5: zone_alpha/hazard_lr interim-clamp constants wired into meta_learn
- B6: overflow_guard::applied_offset() single source of truth
  (workload + RL + device capped at -0.15); metrics_reporter uses it

Mutex / lifecycle (C1/C4, C2, C3, C5, C8):
- C1/C4: battery_low (pct<20 && !charging) disables fast-tick, forces
  min_inter_cycle_ms >= 1000
- C2: reactor cvar uses wait_timeout_while(|t| !*t) — no spurious wake
- C3: daemon_helpers background writer thread (apollo-frozen-writer) for
  frozen_state.json; snapshot under lock, disk I/O off hot path
- C5: FastTickGuard Drop impl clears fast_tick_until on reactor panic/exit
- C8: ActionQueue max_per_cycle getter/setter; daemon halves budget when
  battery pct<30 && !charging

Misc (D2):
- D2: chromium_manager age_cycles u8 -> u16 (u8 saturated at ~128s
  fast-tick); NEW_RENDERER_GRACE_CYCLES typed u16

Deferred:
- C7 lock-ordering invariant comments — structural audit, not a single-
  commit fix. Tracked separately.

cargo check --all-targets: clean (warnings only, all pre-existing pattern).
Tests: frozen_entry_roundtrip + nars_decay_factor assertions updated.

+++ src/engine/chromium_manager.rs
@@ -114,7 +114,12 @@ pub struct RendererInfo {
     frozen_cycles: u8,
     /// Cycles since this renderer was first seen. New renderers (loading a fresh tab)
     /// show 0% CPU before content loads — skip freeze during this warmup window.
-    age_cycles: u8,
+    ///
+    /// D2 fix (round-3): widened `u8 → u16`. With a 500ms fast-tick the u8
+    /// counter saturated at 255 in ~128s, after which age-based LRU thaw
+    /// scoring could not distinguish older-vs-newer frozen renderers and
+    /// the long-idle freeze path was effectively broken for long-lived tabs.
+    age_cycles: u16,
     /// RSS at the moment of freeze. When RSS drops >50% while frozen, the browser
     /// is reclaiming memory from a closed tab — thaw immediately so it can exit cleanly.
     frozen_rss_baseline: u64,
@@ -763,7 +768,7 @@ impl ChromiumManager {
                 // At high pressure (idle_cycles_required=1) it would be frozen before the
                 // page renders, making new tabs appear stuck. Skip freeze for first 10 cycles
                 // (~1s) so the tab has time to load before becoming a freeze candidate.
-                const NEW_RENDERER_GRACE_CYCLES: u8 = 10;
+                const NEW_RENDERER_GRACE_CYCLES: u16 = 10;
                 if info.age_cycles < NEW_RENDERER_GRACE_CYCLES {
                     continue;
                 }
